move to own file and simplify
